Update README to no longer use lines starting with "From"
This is a work-around for an issue with the wheel package, which uses
email.generator to write the METADATA file. The latter, by default, will
write files compatible with Unix mbox format, which means that lines
beginning with "From" are mangled to ">From". This causes the final ReST
document to fail to render on PyPI.
For more information, see:
https://bitbucket.org/pypa/wheel/issues/178/