Fix version handling in 'list_docker_images'
Previously, we listed all images, including those required for older
versions of processes, which caused problems in deployment (e.g. if
an older image was deleted from dockerhub in the meantime).

This commit changes the 'list_docker_images' command to list only
the images required by the latest versions of processes.