Pull Docker images after process registration is complete
This commit adds the '--pull' option to the 'list_docker_images'
management command, as well as a 'post_register_hook' to the
BaseFlowExecutor class, which is called by the 'register' command
after process registration finishes.  This hook is implemented by
the Docker executor to pull Docker images (unless the
'FLOW_DOCKER_DONT_PULL' settings variable is set to True).