Fix race condition when scanning for resolving Data objects
Suppose two manager instances scan the database at roughly the same
time and enter the checking transaction for the same object. Both get to
the end of the transaction at the same time, having checked that the
object is indeed resolving. Both also add it to their own to-do queues
before exiting the transaction. At the end, exactly one of the
transactions will succeed, because they both affect the same rows, so
the other will roll back, however: the object will still stay in that
manager's queue. This means that both managers will proceed to try to
execute this Data object, even though one of them actually had a failed
status update transaction.

The fix is to make adding to the queue dependent on transaction success
by setting a flag in its transaction.on_commit event.