Fix incremental path corruption in the test runner
Commit a5603e59 (part 3/3 of the Channels 2 migration) reorganized the
setup code for parallel testing; this was needed so that the actual
testing call was nested in the call tree below a database_sync_to_async.
An unintended consequence of this was that testing directory path
discovery was done for every batch of tests instead of just as a pool
worker initialization, because Django's RemoteRunner class is
instantiated for each batch by the pool's imap method.

Because the old setup method is still infeasible due to the call nesting
requirements of asyncio, the fix is for the setup functions to now store
state in the runner's testing context, which is process-wide.