Update Preparing release in contributing docs
The Preparing release documentation has been updated to describe the new
workflow where Travis publishes the code to PyPI automatically. We have
required signed commits in the resolwe repository for some time now, so
the documentation was updated accordingly. We have fixed the line breaks
in a few places so that the file is wrapped consistently at 80 (79)
characters.