Retire obsolete processes
We have added the is_active field to the Process model. The field is
read-only on the API and can only be changed through Django ORM. Inactive
processes can not be executed. The register command was extended with the
--retire flag that removes old process versions which do not have
associated data. Then it finds the processes that have been registered but
do not exist in the code anymore, and retires them.