Fix possible infinite loop while deleting data
When persisting error occurred when deleting data for storage location
in storage manager the same location was chosen for processing again
creating an infinite loop.

The status of location is now changed to DELETING and DecisionMaker
consideres only storage locations with status DONE.

Also if same storage location is considered for deleting twice the error
is logged and delete process for file storage instance is stopped.