Fix workflow storage bug
When large data object parent in workflow was processed, its status was
set to DONE before purge started. Status of its StorageLocation object was
set to DONE after purge completed.

If purge was slow it was possible for manager to schedule processing of
another data object (child) in the workflow that depended on parent
before purge of parent data was completed. When that happened the
processing of child object failed with status 'Failed to transfer data.'