Container startup scripts
Python scripts that serve as entrypoints for docker containers.

Processing script starts the processing container and communicates with the
communication contreller via unix domain socket. It requires Python >= 3.4.
This script in injected into container as volume on startup.

Communication script starts the communication container and communicates with
processing container and listener.