Allow upload of large files to S3
The multipart upload to S3 must have no more than 10_000 parts. In order
to upload large files size of each part (part_size) must be such that
the file size is less than 10_000 * part_size.

The computation of s3 etag also depends on the part size so it must be
stored in the database along with the hashes.