Prefer new zmq connections with the same identity
Listener uses zmq router socket for connection handling. When second
client connects with already existing identity, its connection is
rejected and responses are sent to the already connected client.

When zmq.ROUTER_HANDOVER is set to 1 the connection to the already
connected  client is closed and the new one is used.