Copy the settings dictionary to avoid concurency issues
Since multiple commands may be running in separate threads the
settings dictionary must be copied to avoid one thread overriding
settings of the other. Most of the settings are read-only and copy
can be avoided, but RUNTIME_VOLUME_MAPS is set per-process basis
and must be done on a copy of the settings dictionary.