Do not fail in case of missing files in UriResolverView
In UriResolverView, if a pre-signed url was requested for file that
does not exist (or user does not have permission for) an error is
raised. This happens even if only one out of 100 files is missing
or has permission denied.

From now on this is fixed: for failed urls, empty string is
returned and the response is still HTTP 200.